+++ assets/api/helpers.py
@@ -3,9 +3,8 @@ from django.utils.html import format_html
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 
-def generate_button(
-    template, queryset, settings, request
-) -> mark_safe:
+
+def generate_button(template, queryset, settings, request) -> mark_safe:
     """Generate a html button for the tax system"""
     return format_html(
         render_to_string(