+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 - Cache Busting by [@ppfeufer](https://github.com/ppfeufer)
 - Modal System
 - Confirm Modal
-- Single Request Modal
-- View Order Modal
-- Multi Request Modal
+  - View Order Modal
+  - Single Request Modal
+  - Multi Request Modal
 - Request Statistic API
 - Button System
 - RequestAssets Model